Move pkg/netutils to internal/netutils
This change is part of our ongoing effort to move public packages that are not intended for external use into the /internal directory. This prevents accidental external dependencies on our implementation details.

- Moved all files from /pkg/netutils to /internal/netutils
- Updated all import paths throughout the codebase
- No functional changes
