+++ cmd/query/app/server.go
@@ -33,7 +33,7 @@ import (
 	"github.com/jaegertracing/jaeger/internal/recoveryhandler"
 	"github.com/jaegertracing/jaeger/internal/tenancy"
 	"github.com/jaegertracing/jaeger/pkg/bearertoken"
-	"github.com/jaegertracing/jaeger/pkg/netutils"
+	"github.com/jaegertracing/jaeger/internal/netutils"
 	"github.com/jaegertracing/jaeger/pkg/telemetry"
 )
